+++ src/controladores/fastapi/c_cadastrar_usuario.py
@@ -5,7 +5,7 @@ from fastapi import Response, status
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
 
 from src.usecases.erros.erros_usecase import ErroUsuarioExiste
-from src.usecases.uc_cadastrar_usuario import UCCadastrarUsuario
+from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
 from src.usecases.erros.erros_usecase import ErroInesperado
 
 import logging
