+++ src/controladores/fastapi/c_editar_email_fastapi.py
@@ -6,7 +6,7 @@ from devmaua.src.models.erros.erro_usuario import ErroDadosUsuarioInvalidos
 from devmaua.src.models.erros.erro_email import ErroDadosEmailInvalidos
 
 from src.usecases.erros.erros_usecase import ErroInesperado
-from src.usecases.uc_editar_email import UCEditarEmail
+from src.usecases.usuario.uc_editar_email import UCEditarEmail
 
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
