+++ src/controladores/fastapi/c_get_usuario_por_email.py
@@ -1,6 +1,6 @@
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
 from src.usecases.erros.erros_usecase import ErroInesperado
-from src.usecases.uc_get_usuario_por_email import UCGetUsuarioPorEmail
+from src.usecases.usuario.uc_get_usuario_por_email import UCGetUsuarioPorEmail
 from fastapi import Response, status
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
 from fastapi.responses import JSONResponse
