+++ src/controladores/fastapi/c_get_usuario_por_telefone.py
@@ -1,6 +1,6 @@
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
 from src.usecases.erros.erros_usecase import ErroInesperado
-from src.usecases.uc_get_usuario_por_telefone import UCGetUsuarioPorTelefone
+from src.usecases.usuario.uc_get_usuario_por_telefone import UCGetUsuarioPorTelefone
 from fastapi import Response, status
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
 from fastapi.responses import JSONResponse
