+++ src/controladores/fastapi/c_get_usuario_por_userid.py
@@ -1,5 +1,5 @@
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
-from src.usecases.uc_get_usuario_por_userid import UCGetUsuarioPorUserId
+from src.usecases.usuario.uc_get_usuario_por_userid import UCGetUsuarioPorUserId
 from fastapi import Response, status
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
 from src.usecases.erros.erros_usecase import ErroIdInvalido, ErroInesperado
