+++ src/controladores/fastapi/c_remover_endereco_fastapi.py
@@ -6,7 +6,7 @@ from devmaua.src.models.erros.erro_usuario import ErroDadosUsuarioInvalidos
 from devmaua.src.models.erros.erro_endereco import ErroDadosEnderecoInvalidos
 
 from src.usecases.erros.erros_usecase import ErroInesperado
-from src.usecases.uc_remover_endereco import UCRemoverEndereco
+from src.usecases.usuario.uc_remover_endereco import UCRemoverEndereco
 
 from src.interfaces.IRepoUsuario import IArmazenamentoUsuario
