+++ src/usecases/usuario/uc_get_usuario_por_telefone.py
