+++ tests/controladores/test_c_editar_endereco_fastapi.py
@@ -1,6 +1,6 @@
 from src.controladores.fastapi.c_editar_endereco_fastapi import ControllerHTTPEditarEnderecoFastAPI
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
-from src.usecases.uc_cadastrar_usuario import UCCadastrarUsuario
+from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
 
 from devmaua.src.models.usuario import Usuario
