+++ tests/usecases/usuario/test_cadastrar_usuario.py
@@ -10,7 +10,7 @@ from devmaua.src.models.telefone import Telefone
 from devmaua.src.models.usuario import Usuario
 
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
-from src.usecases.uc_cadastrar_usuario import UCCadastrarUsuario
+from src.usecases.usuario.uc_cadastrar_usuario import UCCadastrarUsuario
 from src.usecases.erros.erros_usecase import ErroUsuarioExiste
