+++ tests/usecases/usuario/test_uc_get_usuario_por_email.py
@@ -4,7 +4,7 @@ from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsu
 from devmaua.src.models.usuario import Usuario
 import tests.mock_objetos as mo
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
-from src.usecases.uc_get_usuario_por_email import UCGetUsuarioPorEmail
+from src.usecases.usuario.uc_get_usuario_por_email import UCGetUsuarioPorEmail
 
 
 class TestUCGetPorEmail:
